authorization completed || start auth for check admin to access admin/panel

+++ app/Http/Controllers/Admin/FileController.php
@@ -4,14 +4,18 @@ namespace App\Http\Controllers\Admin;
 
 use App\File;
 use App\Http\Controllers\Controller;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\View\View;
 
 class FileController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
@@ -22,7 +26,7 @@ class FileController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
@@ -32,8 +36,8 @@ class FileController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
     public function store(Request $request)
     {
@@ -45,7 +49,7 @@ class FileController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function show($id)
     {
@@ -55,20 +59,24 @@ class FileController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  File  $file
-     * @return \Illuminate\Http\Response
+     * @param File $file
+     * @return Factory|View
+     * @throws AuthorizationException
      */
     public function edit($file)
     {
+        //if you are agent, you can't access to this route
+        $this->authorize('admin_permissions');
+
         return view('Admin.files.edit', compact('file'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param int $id
+     * @return void
      */
     public function update(Request $request, $id)
     {
@@ -78,11 +86,14 @@ class FileController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  File  $file
-     * @return \Illuminate\Http\Response
+     * @param File $file
+     * @return Response
+     * @throws AuthorizationException
      */
     public function destroy($file)
     {
+        $this->authorize('admin_permissions');
+
         $file->delete();
         return back();
     }
